+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 Also this project likes fast realizes with even one change to get it live shortly.
 
+## 1.0.0 - 2018-12-21
+- Add `options` as last argument to all functions making queries.
+- Replace `transaction` function argument with `options`.
+
 ## 0.1.0 - 2018-12-19
 - Replace arrow functions to 'function'.
 - Add `lft` and `rgt` to `getDescendants` function.